Application SAT Solver (Part 2)
Part 1 detected where multiple dependencies affected version selection
and combined them to choose one, if any, that satisfied those
constraints.  What it didn't do was handle when a version has already
been selected and a new constraint was not met.  This detects that
condition and rolls back to the previous descision and tries with a new
version, repeating the process until the problem is satisfied or not.